fix logging.
because we were closing/reopening the file with every logging attempt,
we ended up writing half of it to one dir, and half to the other.

Things will be a lot simpler if we just leave the file open.
The fflush/fsync should ensure that things still hit the disk.