beginning of parent/child state machine
The basic idea is that we now always keep n threads busy.
Previously the parent would wait for the last thread to exit
before spawning new ones. Now as soon as one dies, a new one
takes its place.

I had a version of this that worked better earlier this evening,
but didn't commit it before I accidentally trashed up my source tree.
In some ways this attempt is cleaner, but it's in need of some
more debugging, which I'll attempt with a clearer head tomorrow.