avoid ptrace for now.
This is going to need some more thought.
There are a number of awkward cases where we end up with zombies.
One example:

child a ptraces child b
child a segfaults
child b waits forever for a PTRACE_DETACH/CONT

attempting a PTRACE_DETACH from the watchdog gets -ESRCH, because
it wasn't the pid that initiated the tracing.