add some sanity checking to the shm.
Very rarely, there's a case where we get nonsense pids in the shm.
I've seen it happen twice now, so add some debugging to catch it
so it can be properly debugged.

If you see this happen, attach to a trinity process with gdb,
and print out the values of shm->previous_syscallno, and shm->previous_a[1-6]
Hopefully that will show the syscall that's doing the corrupting.