timeval wrap workaround
if a thread completes a syscall after we sample 'old', we can appear
to have wrapped. Most the time, this shows up as a single second delta,
but on occasion if the system is heavily loaded it can take a few more.
This doesn't fix the problem entirely, but it should lessen it.