Scale down struct arglist size
Scale down the struct arglist values array size. The most we are using
at the moment is 23 entries from the array in a few syscalls, so 1024 is
way excessive. It also bloats the resulting binary size: after this
patch the stripped trinity binary size went from ~10 megabytes to 500
kilobytes on my system.

GCC also gives a nice warning in case the array size limit is exceeded.