Revert "Fix -c to only call the syscall specified."
This reverts commit 819c566ed0c277b8ce80267da12a8ced3e2b8b00.

Turns out that wasn't quite so dumb after all.
Sasha points out that I did it this way so that multiple -c parameters
can be passed.  With the commit above, we could only have a single -c.

To do this properly we'd have to construct a special syscall table,
and then change the rand() to be nr_enabled_syscalls.

For now, just revert. I'll revisit this post 1.1 sometime.