overwritten assignment in sanitise_setsockopt
I noticed an assignment that gets immediately overwritten, contradicting
a nearby comment.
According to the comment, 'optval' should get disabled half of the time.

As at the time of the conditional assignment optval was already set to
different nonzero values, it looks to me like the trailing assignment,
which was introduced in commit 737e1a71, may be unintended.

Signed-off-by: Nicolas Kaiser <nikai@nikai.net>