Add support for cross building using CROSS_COMPILE
Similarly to the kernel.  CROSS_COMPILE is optional, when specified it's
a prefix, eg. myarch-linux-gnu-

We still allow overriding CC, separately, or in addition to setting
CROSS_COMPILE.

There is also support for cross configuring, because at the moment all
the configure tests only need to build, they're not run.