add init() function to struct syscall
While working on the perf_event_open() sysfs init problem, I was wondering
if it might be easier if we added the possibility of an init() routine
to each syscall structure.  That way trinity can support doing setup
before fuzzing begins.

Below is a quick patch implementing this, it seems to work but I still
don't have the best grasp of trinity internals.

Signed-off-by: Vince Weaver <vincent.weaver@maine.edu>