Split shm->parentpid into two.
In some cases this was 'initial process', and the rest of the time 'forking thread'.
Split into two so watchdog monitors only the forking process.

This way, if for whatever reason the forking thread dies (usually a segv)
then we don't hang, because the watchdog will see it's no longer around
instead of it checking on the init process.

Additionally, these pids don't need to be in the shm.