Added parameter that controls which taint flag causes trinity to stop.
This is reworked patch for the parameter that allows to control which
kernel taint flags causes trinity to stop. In particular, now the short
parameter is -T (instead of -t) and flags has to be specified as literals
(e.g., DIE, WARN, USER, etc.). All 13 supported flags are specified in
README.

Signed-off-by: Ildar Muslukhov <ildarm@google.com>