save/restore r10 around 32bit x86-64 syscalls.
This fixes the problem where we segfault as soon as we hit __syscall_return(res)
res was being held in r10, and went out of scope once we hit the syscall,
so we were scribbling through random ptrs ->segv.

We should probably be saving more registers here to be completely safe,
but for now this seems to do the trick.