remove logging usage of previous_a*
This is bogus. The old code compared an unaltered version of the register
to the post-sanitised version.  This 'oldreg' bears no relation to
shm->previous_a* which is the _previous syscall_ that was made.
The latter is a debug thing, not something useful in this way.

We only did this to colorize the output some, so it's simpler just
to remove that whole thing.