Fixes bug with incorrect casting of on stackvariable to a pointer.
This patch fixes a bug in open_socket function with sa variable of type
struct sockaddr. In particular, this variable is alloctaed on stack, and
then address to it is passed to specific functions that allocate memory
and overwrite the content of sa with the address of the newly allocated
object, but not with the values that were assigned to the object. Thus, 4
or 8 first bytes of the sa struct are always initialized with an
address of newly allocated object.

Signed-off-by: Ildar Muslukhov <ildarm@google.com>