split up shm create again.
Now I remember why we had that _postargs function.
The params parsing toggles shm->syscall* so it needs the shm set up
before we do the param parsing.
And we need to have done the param parsing before we do the array
allocation in case we've passed -C.

The answer is to do this in two phases.