make perf_event_open() create system-wide events
This patch adds a create_mostly_valid_global_event case to the
perf_event_open() code.

Without this patch it is extremely unlikely that a global
(not per-process) event would be randomly created.

Most users will not notice a change, as you can only create global events
as root or else if /proc/sys/kernel/perf_event_paranoid is set to < 1
(the default is 1).

With this change some of the more obscure event types such as uncore,
northbridge, and RAPL power events are occasionally properly created.

Signed-off-by: Vince Weaver <vincent.weaver@maine.edu>