Make fork failure only fatal if we have no running children.
This is an experiment, and one I tried before, but let's see if
we can at least do better than just exiting the first time fork() fails.

Last time I tried this, I ended up with a ton of fork failures,
so we might still need to do something else but that was before
I did a lot of the VM work, so things may be better now anyway.