remove the outputting of the reseeding.
This causes a *lot* of the spew from trinity, and will only get worse
as child processes start segv'ing more often.

now that the watchdog dumps the last known seed when it detects taint,
that is hopefully good enough.

The one corner case is where we crash so badly that the watchdog
doesn't get time to write the last seed to syslog.  In that
circumstance, we used to be able to look back at the last known
good seed.

A better solution might be to dump some state periodically over
the network to another system.  I'll experiment with this at some point.