make output_syscall_prefix_to_fd use get_syscall_entry
This way we work out the correct table to read from.
In doing this conversion I realised that the syscall rec only has the
actual syscall number in it, and we need the offset here.

One way to do this might be to make get_syscall_entry not care,
by having it have knowledge of SYSCALL_OFFSET. needs more thinking.