comment out the post mortem code again.
This is going to need a total reworking, as this way of doing things
isn't going to work.  If a child does something like allocating
a page to mangle a pathname, we can't output that pathname from
the watchdog context because the allocation is child local.
So right now, we pass invalid ptrs to print as %s, and things segv.

Not sure what the solution looks like, but this isn't it.