increase the prebuffer size.
We have potential here for buffer overflows.
Imagine for eg, we have a syscall that needs a pathname as an arg,
and we end up allocating a 4k buffer of garbage to pass.
The output routines try to render that 4k and happily scribble past
the arrays.  They should probably be more mindful of the lengths
being passed in so that they don't overrun, but for now just increase
the size of the arrays. We do it * 6, because in theory we could pass
junk pages like this in every register (though unlikely in reality).