Fix powerpc build on Ubuntu
In commit df1508cd4151 "move rtas syscall definition to ppc specific
directory" the rtas syscall got moved to syscalls/ppc, and we added a
rule in the Makefile to pick that up based on the output of
gcc -dumpmachine.

That rule only looks for "ppc*" though, on Ubuntu gcc emits
"powerpc-linux-gnu", or "powerpc64le-linux-gnu".