configure.sh: Respect CFLAGS from the environment
Currently, configure.sh does not inherit any cflags from the
environment. However, for MIPS, it's necessary to pass additional
cflags to decide, for example, if we want a big endian (-BE) or
little endian (-EL) build. As a result of which, we modify the
configure.sh script to respect additional cflags passed by the
user.

Signed-off-by: Markos Chandras <markos.chandras@imgtec.com>