Revert "allocate the output buffers on child startup."
This reverts commit 10e37a39ed8686cb1801adc42bfbb67ce39975ce.

While this had the effect of avoiding the occasional corruption
to the child structs, it had the unwanted side-effect of breaking
the post-mortem code, because it could no longer dereference
the buffers from outside the child process.

I still need to spend some time hardening the output functions
to not overrun these buffers, but since 1280b9854b, we should
at least not hang when we hit corrupted states any more, which
is a good enough stop-gap until I fix things up properly.