remove the taviso random function
This came from an era when I was trying to figure out why iknowthis
was finding bugs that trinity wasn't. It turned out to not be
because of the different random function, but because we kept
hitting other bugs first due to the differences in our random
generation. Fix the other bugs, and run trinity for long enough,
and you'd also hit the iknowthis bugs.

(There were also a few ->sanitize improvements for some syscalls
 along the way).

Either way, this hasn't actually been useful for over a year, and
just adds pointless complexity to the random code.