futex: pimp futex fuzzing
Make futex fuzzing more ad-hoc and spice things up to improve
testing coverage.  Roughly half the calls will use the generic
arguments, with the occasional exception of using CLOCK_REALTIME,
when applicable to the correct cmd (avoid -ENOSYS). The other half
can perform anyone of the following operations, each with a ~25%
chance, acting on at most two uaddresses from a small pool selected
randomly.
 - Perform a pseudo trylock/unlock. Contended scenarios
    involve immediately blocking or doing a pi lock.

 - Wait/requeue, which pairs with one of the below wakes.
 - WAKE_OP, or,
 - a regular wake or directly requeueing.

Additionally, update the TODO file as futexes are not (de)initialized,
and we can just use the handful provided locality by futex.c. These are
an array of locks which are used for the corresponding uaddresses requred
by the system call.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>