futex: convert to object cache
This allows better handling of the uaddr parameters we pass
to futex(2). As such, teach the object structure about the
internal lock primitive. We have two alternatives now, which
replaces the current use lock or whatever trinity gives us
as addresses. Now we continue using the lock half the time,
but make use of the mmap playground already available otherwise.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>