check for empty pidslots in stall_genocide
This was a pain in the ass to track down. We were sigkill'ing
a pid of -1, which took down everything.
Amazingly, this has been there for six months, and only recently
became a problem.  This is because children previously weren't
running long enough, or getting stuck enough for this situation
to arise that often.  With the recent segfault fixes, this
function gets hit a lot more.