futex: fix uaddr passing when dealing with lock
We are working on the address of the futex word in our
local lock structure. As such, we are currently passing
it a completely bogus params for uaddr and uaddr2, which
can be seen with the amount of unsuccessful futex calls.

Also get rid of printing the addresses in the mmap'ed
playground when creating our futexes, just show how many
are created, no one really cares about these specifics,
which can be seen/studied in the logs anyway.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>