Remove the watchdog process.
Have the work it used to do done by the main process instead.
There's still room for improvement here, but this is a big step
forward in terms of simplification, and allows a whole bunch of
code & locking to just go away.

There's also no longer a good reason to fork main_loop() as a separate
process, so now we're back to a single process (for now).

A big change here also is the 'reaping' of processes. This is now
only ever done by the main process. The child no longer cleans up
after itself (ugh, kids eh?) before it exits.  There's no watchdog
process, so we can probably remove the reaper_lock now as a follow-up.