Add MCL_ONFAULT to mlockall calls
The new mlock2 system call was added, but it looks like the MCL_ONFAULT
flag was missed.  This is now a valid flag to pass to mlockall(), but it
must be used in conjunction with either MCL_CURRENT or MCL_FUTURE (or
both).  Also rework the way that the flags value for mlockall() is
caclulated.  It is valid to pass any combination of the MCL_* flags
except for MCL_ONFAULT alone.

Signed-off-by: Eric B Munson <emunson@akamai.com>
Cc: Dave Jones <davej@codemonkey.org.uk>