change check for inactive children
Jan Stancek reported that he saw processes hanging without getting killed,
and this is the most probably reason.  We were considering children
as inactive even if they were doing the syscall, so they were never getting
sent a SIGKILL.

Change the state comparison, and also widen the window that we spend
in 'BEFORE' up until right before the syscall.