Fix version.h generation
The makefile doesn't specify correct dependencies for the version
header and as a result we may end up compiling sources which
require the header before the header is generated.

Rewrite the dependencies so that:
- version.h has a dependency on all of the files that may change it
- all .c files have an order-only dependency on version.h

An order-only dependency on version.h should be enough for all files,
as any sources that have a direct dependency on the version.h will
later use the normal dependencies defined on .d files. Those, however,
don't exist on the first make run so this extra rule is required.

Finally, update gen-versionh.sh to touch the file even if it hasn't
changed, so that if any pre-requisites change in a way that trigger
version.h regeneration, we still update its timestamp to keep make
happy.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@imgtec.com>