syscall: ensure a child is not killed with a lock held
When an EXTRA_FORK syscall is executed, there is a race condition
between the forked child and the parent - one trying to execute
the syscall and the other trying to kill it after a short delay.

It is therefore possible for the child to be killed while it's
holding a syscallrecord lock. Prevent this by taking the lock
before attempting to kill the child.

Theoretically this may lead to a situation when the parent takes
the lock before the child completes preventing it from updating
the record state to AFTER and setting the return value. However,
since this is for an EXTRA_FORK case only where we expect the
syscall to never return and get replaced by the new process, it
is not an issue as the post-syscall record update sequence is
never reached.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@imgtec.com>