Add SIGILL as a potential child exit signal
It is possible for the child process to terminate with a SIGILL signal.
As this is currently not handled, the parent process would not properly
clean up the child's state and will later assume the child is not making
a progress and attempt to kill it even though it's no longer there.

Signed-off-by: Marcin Nowakowski <marcin.nowakowski@imgtec.com>