ioctls/kvm: fix compile error
v6.9-rc1 introduced some kvm uapi header structure reorganization:
  71cd774ad2f KVM: s390: move s390-specific structs to uapi/asm/kvm.h
  d750951c9ed KVM: powerpc: move powerpc-specific structs to uapi/asm/kvm.h
v6.8-rc1:
  a5d3df8ae13 KVM: remove deprecated UAPIs

Arch related files are moved to arch folders. And some deprecated
macros are delated. so update the kvm ioctl to check that, otherwise
it would fail to compile.

Use KVM_TRC_SHIFT to check if the deprecated macros are deleted, if
it's not there, then deprecated macros are deleted.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>