kvm: move all s390 ioctls inside ifdef
Linux 6.9 cleaned up some of the struct definitions for architecture-specific
KVM ioctls, moving them to the headers for that architecture in asm/.  However,
the result is that ioctls are defined _but_ not usable.  This can be fixed
upstream, but is it worth it?  There are actually a lot of other cases where
the ioctls were placed in linux/kvm.h (to avoid numbering conflicts) but
the struct is only defined for specific architectures, so place
these s390-specific ioctls in the same group.

Reported-by: Cédric Le Goater <clg@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>