ioctl/drm: fix drm_nouveau compile issue (#53)
since 460be1d527a8 ("drm/nouveau: move more missing UAPI bits") in
v6.10-rc1, the two structs are moved to the uapi in
include/drm/drm_nouveau.h (drm_nouveau_notifierobj_alloc and
drm_nouveau_gpuobj_free), so the definition of the macros and the
structs are not in the same place any more, so can't use the macros
to determine if the two structs are defined. So change to compile in
configure to determine that.